+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ v4.1.1 (UNRELEASED)
 -------------------
 * :star: We added the ProjectInfo widget.
 * :star: We added the concept of prefilled_parts that are stored on the Form. These prefilled parts are instance that are created on form instantiation. This in combination with grid like forms in any Form.
+* :star: We added helpers to manage the alteration of the project_info attribute on the Scope object better.
 
 v4.1.0 (24MAY22)
 ----------------